refactor(azure-storage): make the Deep Agents backend a public subpackage
Rename the internal _deepagents package to a public
langchain_azure_storage.deepagents and import the backend from there
(from langchain_azure_storage.deepagents import AzureBlobBackend).

Drop the top-level PEP 562 __getattr__ re-export and move the
"install the [deepagents] extra" ImportError into the subpackage's
__init__, so the dependency requirement surfaces clearly on first
import. Also trim the internal _user_agent module docstring.
